Chore: Move bulk-add judges/teams to an RPE functionality
This will move the bulk-add judges/teams to an RPE functionality to the
new `/concerns/regional_pitch_events/` folder.

Refs: #5941
